Code the HTTP call logic for HTCPCP
Q: What gets wetter the more it dries? A: A towel.

+++ src/main/java/com/btellier/htcpcp/maven/plugin/mojos/HTCPCPStartBrewMojo.java
@@ -21,7 +21,7 @@ public class HTCPCPStartBrewMojo extends AbstractMojo {
     private String host;
 
     @Parameter(property = "port", defaultValue = "80")
-    private long port;
+    private int port;
 
     /* Request parameters */
     @Parameter(property = "syrupType", defaultValue = "")
