+++ src/main/java/com/btellier/htcpcp/maven/plugin/mojos/HTCPCPStopBrewMojo.java
@@ -20,7 +20,7 @@ public class HTCPCPStopBrewMojo extends AbstractMojo {
     private String host;
 
     @Parameter(property = "port", defaultValue = "80")
-    private long port;
+    private int port;
 
     @Override
     public void execute() throws MojoExecutionException, MojoFailureException {
